+++ project/at_start_f403a/examples/usb_device/msc/inc/msc_diskio.h
@@ -51,8 +51,8 @@ extern "C" {
 #define SECTOR_SIZE_4K                   4096
 
 uint8_t *get_inquiry(uint8_t lun);
-usb_sts_type msc_disk_read(uint8_t lun, uint32_t addr, uint8_t *read_buf, uint32_t len);
-usb_sts_type msc_disk_write(uint8_t lun, uint32_t addr, uint8_t *buf, uint32_t len);
+usb_sts_type msc_disk_read(uint8_t lun, uint64_t addr, uint8_t *read_buf, uint32_t len);
+usb_sts_type msc_disk_write(uint8_t lun, uint64_t addr, uint8_t *buf, uint32_t len);
 usb_sts_type msc_disk_capacity(uint8_t lun, uint32_t *blk_nbr, uint32_t *blk_size);
 
 /**
