+++ project/at_start_f407/examples/usb_device/virtual_msc_iap/src/msc_diskio.c
@@ -70,13 +70,13 @@ uint8_t *get_inquiry(uint8_t lun)
   * @param  len: read length
   * @retval status of usb_sts_type
   */
-usb_sts_type msc_disk_read(uint8_t lun, uint32_t addr, uint8_t *read_buf, uint32_t len)
+usb_sts_type msc_disk_read(uint8_t lun, uint64_t addr, uint8_t *read_buf, uint32_t len)
 {
 
   switch(lun)
   {
     case 0:
-      flash_fat16_read(addr, read_buf, len);
+      flash_fat16_read((uint32_t)addr, read_buf, len);
       break;
     case 1:
       break;
@@ -92,12 +92,12 @@ usb_sts_type msc_disk_read(uint8_t lun, uint32_t addr, uint8_t *read_buf, uint32
   * @param  len: write length
   * @retval status of usb_sts_type
   */
-usb_sts_type msc_disk_write(uint8_t lun, uint32_t addr, uint8_t *buf, uint32_t len)
+usb_sts_type msc_disk_write(uint8_t lun, uint64_t addr, uint8_t *buf, uint32_t len)
 {
   switch(lun)
   {
     case 0:
-      flash_fat16_write(addr, buf, len);
+      flash_fat16_write((uint32_t)addr, buf, len);
       break;
     case 1:
       break;
